requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
asgiref==3.7.2
aspose-words==23.9.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.1.0
black==23.9.1
blinker==1.6.2
Brotli==1.1.0
brotlipy==0.7.0
cachetools==5.3.1
camelot @ file:///home/ubuntu/draft/writer/openai_skt/database/chunk/camelot
camelot-py==0.9.0
celery==5.3.4
certifi==2023.7.22
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.8
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==2.2.1
cmake==3.27.5
coloredlogs==15.0.1
commonmark==0.9.1

configparser==6.0.0
contourpy==1.1.1
cryptography==41.0.4
cssselect2==0.7.0
cycler==0.11.0
dataclasses-json==0.5.14
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@8c4a333ceb8df05348759443d0206302485890e0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
drf-yasg==1.21.7
elastic-transport==8.4.0
elasticsearch==8.9.0
embedchain==0.0.54
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.99.1
filelock==3.12.4
Flask==2.3.3
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fvcore==0.1.5.post20221221
ghostscript==0.7
gobject==0.1.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.58.0
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
httptools==0.6.0
humanfriendly==10.0
hydra-core==1.3.2

importlib-resources==6.0.1
inflection==0.5.1
iopath==0.1.9
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.1
kiwisolver==1.4.5
kombu==5.3.2
langchain==0.0.279
langsmith==0.0.35
layoutparser==0.3.4

lit==16.0.6
lxml==4.9.3
Markdown==3.4.4
markdown2==2.4.10
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
md2pdf==1.0.1
mdpdf==0.0.18
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
nest-asyncio==1.5.7
networkx==3.1
numexpr==2.8.5
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.15.1
openai==0.27.10
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.76
openpyxl==3.1.2
outcome==1.2.0
overrides==7.4.0
packaging==23.1
pandas==1.5.3
pango==0.0.1
pathspec==0.11.2
pdf2image==1.16.3
pdfminer==20191125
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.1
platformdirs==3.10.0

portalocker==2.8.2
posthog==3.0.2
prompt-toolkit==3.0.39
protobuf==4.24.2
pulsar-client==3.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7

pycryptodome==3.19.0
pydantic==1.10.12
pydantic_core==2.6.3
pydyf==0.7.0
PyJWT==2.8.0
PyMuPDF==1.23.3
PyMuPDFb==1.23.3

pyparsing==3.1.1
pypdf==3.15.4
PyPDF2==3.0.1
pypdfium2==4.20.0
pyphen==0.14.0
PyPika==0.48.9

python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rsa==4.9

scikit-learn==1.3.1
scipy==1.11.2
selenium==4.12.0

sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
sqlparse==0.4.4
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.2.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
triton==2.0.0
twilio==8.7.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.3.7
wsproto==1.2.0
yacs==0.1.8
yarl==1.9.2
youtube-transcript-api==0.6.1
zopfli==0.2.3
